ADD: overloaded generate method and made it using float parameters by default

+++ src/main/java/pl/admonster/ClassInstancesGenerator/service/RandomNumeric.java
@@ -9,7 +9,7 @@ public class RandomNumeric {
     }
 
     static float generate(int minRange, int maxRange) {
-        return generate(minRange, maxRange);
+        return generate((float) minRange, (float) maxRange);
     }
 
     static float generate(NumericValuePrototype prototype) {
